Moved get_bool to apps.tools.string
